+++ tests/manual/catalog_test.js
@@ -1,7 +1,7 @@
 // const Catalog_Upsert = require("../../src/lib/catalog_request_upsert");
 // const sample_objects = require("../data_preparation/sample_catalog_data");
 const Catalog_List = require("../../src/lib/catalog_request_list");
-const Catalog_Search_Filter = require("../../src/lib/stub.catalog_request_search_objects_filter");
+const Catalog_Search_Filter = require("../../src/lib/catalog_request_search_objects_filter");
 
 const list = new Catalog_List();
